Moved server components to `@/server/`
